Remove Ray from integration tests; replace with Iris backend
- Replace tests/integration_test.py with the Iris-backed version
  (previously tests/integration/iris/run_iris_full_integration.py),
  removing the ray.init() + RayClient startup entirely.
- Delete the now-redundant iris integration script.
- Update marin-itest.yaml: drop the "Give Ray tmp space" step and
  RAY_TMPDIR env var; gate the job on workflow_dispatch with a
  --controller-url input (requires an Iris cluster).
- Remove ray_cluster fixture and "ray" cluster/job_context params
  from lib/fray/tests/conftest.py and test_job_context.py.
- Delete lib/fray/tests/test_v2_ray.py and test_ray_tpu.py.
