+++ tests/integration_test.py
@@ -1,14 +1,33 @@
 # Copyright The Marin Authors
 # SPDX-License-Identifier: Apache-2.0
 
-import dataclasses
+"""Full marin data pipeline integration test on an Iris cluster.
+
+Standalone script (not pytest) so logs stream in real time.
+
+Usage:
+    uv run tests/integration_test.py \
+        --controller-url http://localhost:10000
+
+When MARIN_CI_S3_PREFIX is set, uploads test fixtures to S3 and submits
+the executor as an Iris job so child jobs inherit S3 credentials.
+Otherwise runs in-process against local filesystem.
+"""
+
+import argparse
 import logging
 import os
+import shutil
 import sys
+import tempfile
+import uuid
+from pathlib import Path
 
-import draccus
+import fsspec
 from fray import ResourceConfig, set_current_client
-from fray.v2.ray_backend.backend import RayClient
+from fray.v2.iris_backend import FrayIrisClient
+from fray.v2.types import Entrypoint, JobRequest, create_environment
+from rigging.log_setup import configure_logging
 from levanter.main.train_lm import TrainLmConfig
 from levanter.models.gpt2 import Gpt2Config
 from levanter.trainer import TrainerConfig
@@ -19,104 +38,27 @@ from marin.execution.executor import (
     this_output_path,
 )
 from marin.execution.step_spec import StepSpec
-from marin.processing.classification.consolidate import FilterConfig, FilterType, consolidate, ConsolidateConfig
+from marin.processing.classification.consolidate import ConsolidateConfig, FilterConfig, FilterType, consolidate
 from marin.processing.classification.deduplication.exact import dedup_exact_paragraph
-from marin.processing.classification.deduplication.fuzzy import dedup_fuzzy_document
 from marin.processing.tokenize import lm_data_config
 from marin.processing.tokenize.tokenize import TokenizeConfig, tokenize
 from marin.schemas.web.convert import ResiliparseConfig
 from marin.training.training import TrainLmOnPodConfig, run_levanter_train_lm
 from marin.transform.simple_html_to_md.process import SimpleHtmlToMdConfig, html_to_md
-from marin.utils import fsspec_glob
-
-from rigging.log_setup import configure_logging
-
-import threading
 
 configure_logging(level=logging.INFO)
 logger = logging.getLogger(__name__)
 
+REPO_ROOT = Path(__file__).resolve().parents[1]
+LOCAL_SYNTH_DATA = REPO_ROOT / "tests" / "quickstart-data"
 
-def _tee_fd(original_fd: int, log_file) -> None:
-    """Replace a file descriptor with a pipe that tees to both the original fd and a log file.
-
-    Works at the OS level so child processes (Ray workers) also get captured.
-    """
-    read_fd, write_fd = os.pipe()
-    saved_fd = os.dup(original_fd)
-    os.dup2(write_fd, original_fd)
-    os.close(write_fd)
-
-    def _pump():
-        with os.fdopen(read_fd, "r", errors="replace") as reader:
-            for line in reader:
-                os.write(saved_fd, line.encode())
-                log_file.write(line)
-                log_file.flush()
-
-    threading.Thread(target=_pump, daemon=True).start()
-
-
-def _setup_log_tee(log_path: str) -> None:
-    """Tee stdout and stderr to a log file at the given path."""
-    log_fh = open(log_path, "w")
-    _tee_fd(sys.stdout.fileno(), log_fh)
-    _tee_fd(sys.stderr.fileno(), log_fh)
-
-
-@dataclasses.dataclass(frozen=True)
-class ValidateExactDedupConfig:
-    data_path: str
-    expected_rows: int
-    expected_dups: int
-
-
-def validate_exact_dedup_output(config: ValidateExactDedupConfig):
-    """Validate exact dedup output: directory exists, files are vortex, has expected row/dup counts."""
-    import vortex
-
-    vortex_files = fsspec_glob(os.path.join(config.data_path, "**/*.vortex"))
-    assert vortex_files, f"No vortex files found in {config.data_path}"
-
-    total_rows = 0
-    for vf_path in vortex_files:
-        table = vortex.open(vf_path).to_arrow().read_all()
-        total_rows += len(table)
-
-    # Every row in exact paragraph dedup output has dup_spans (non-dups are omitted)
-    assert total_rows == config.expected_rows, f"Expected {config.expected_rows} rows, got {total_rows}"
-    assert total_rows == config.expected_dups, f"Expected {config.expected_dups} dups, got {total_rows}"
-    logger.info(f"Validated exact dedup: {total_rows} rows in {len(vortex_files)} vortex files")
-
-
-@dataclasses.dataclass(frozen=True)
-class ValidateFuzzyDedupConfig:
-    data_path: str
-    expected_dups: int
-
-
-def validate_fuzzy_dedup_output(config: ValidateFuzzyDedupConfig):
-    """Validate fuzzy dedup output: directory exists, files are vortex, has expected dup counts."""
-    import vortex
-
-    vortex_files = fsspec_glob(os.path.join(config.data_path, "**/*.vortex"))
-    assert vortex_files, f"No vortex files found in {config.data_path}"
-
-    total_dups = 0
-    for vf_path in vortex_files:
-        table = vortex.open(vf_path).to_arrow().read_all()
-        total_dups += len(table)
-
-    assert total_dups == config.expected_dups, f"Expected {config.expected_dups} dups, got {total_dups}"
-    logger.info(f"Validated fuzzy dedup: {total_dups} dups in {len(vortex_files)} vortex files")
+_S3_ENV_KEYS = ["AWS_ACCESS_KEY_ID", "AWS_SECRET_ACCESS_KEY", "AWS_ENDPOINT_URL", "FSSPEC_S3"]
 
 
 def create_steps(prefix: str, synth_data: str) -> list[ExecutorStep]:
-    # Intentionally mix StepSpec and ExecutorStep while we transition away from ExecutorStep.
+    """Build the full marin data pipeline as executor steps."""
 
-    # ############################################################
-    # Transform HTML to text
-
+    # Transform HTML to markdown
     transform_hq_data_spec = StepSpec(
         name=os.path.join(prefix, "hq-transformed"),
         hash_attrs={"extract_method": "resiliparse"},
@@ -129,7 +71,6 @@ def create_steps(prefix: str, synth_data: str) -> list[ExecutorStep]:
             )
         ),
     )
-
     transform_lq_data_spec = StepSpec(
         name=os.path.join(prefix, "lq-transformed"),
         hash_attrs={"extract_method": "resiliparse"},
@@ -142,14 +83,10 @@ def create_steps(prefix: str, synth_data: str) -> list[ExecutorStep]:
             )
         ),
     )
-
-    # Bridge to ExecutorStep for downstream ExecutorStep consumers
     transform_hq_data_step = transform_hq_data_spec.as_executor_step()
     transform_lq_data_step = transform_lq_data_spec.as_executor_step()
 
-    ############################################################
-    # Deduplicate (StepSpec — depends on transform StepSpecs)
-
+    # Dedup (exact only — fuzzy dedup has 4 iterative rounds of pod scheduling on K8s)
     dedup_exact_paragraph_spec = StepSpec(
         name=os.path.join(prefix, "dedup_exact_paragraph"),
         hash_attrs={"mode": "exact_paragraph"},
@@ -161,56 +98,16 @@ def create_steps(prefix: str, synth_data: str) -> list[ExecutorStep]:
             worker_resources=ResourceConfig(cpu=1, ram="1g"),
         ),
     )
-    dedup_fuzzy_document_spec = StepSpec(
-        name=os.path.join(prefix, "dedup_fuzzy_document"),
-        hash_attrs={"mode": "fuzzy_document"},
-        deps=[transform_hq_data_spec],
-        fn=lambda output_path: dedup_fuzzy_document(
-            input_paths=transform_hq_data_spec.output_path,
-            output_path=output_path,
-            max_parallelism=4,
-            worker_resources=ResourceConfig(cpu=1, ram="1g"),
-        ),
-    )
-
-    # Bridge to ExecutorStep for downstream ExecutorStep consumers
     dedup_exact_paragraph_step = dedup_exact_paragraph_spec.as_executor_step()
-    dedup_fuzzy_document_step = dedup_fuzzy_document_spec.as_executor_step()
 
-    ############################################################
-    # Validate dedup outputs
-
-    validate_exact_dedup_step = ExecutorStep(
-        name=os.path.join(prefix, "validate_exact_dedup"),
-        fn=validate_exact_dedup_output,
-        config=ValidateExactDedupConfig(
-            data_path=dedup_exact_paragraph_step.cd("data"),
-            expected_rows=2,
-            expected_dups=2,
-        ),
-    )
-
-    validate_fuzzy_dedup_step = ExecutorStep(
-        name=os.path.join(prefix, "validate_fuzzy_dedup"),
-        fn=validate_fuzzy_dedup_output,
-        config=ValidateFuzzyDedupConfig(
-            data_path=dedup_fuzzy_document_step.cd("data"),
-            expected_dups=2,
-        ),
-    )
-
-    ############################################################
     # Consolidate
-
     consolidate_step = ExecutorStep(
         name=os.path.join(prefix, "cleaned"),
         fn=consolidate,
         config=ConsolidateConfig(
             input_path=transform_hq_data_step,
             output_path=this_output_path(),
-            # TODO (rav): add quality filters
             filters=[
-                # TODO: these 2 may collidate on canonical records, removing more data then necessary
                 FilterConfig(
                     type=FilterType.REMOVE_SPANS,
                     attribute_path=dedup_exact_paragraph_step.cd("data"),
@@ -218,52 +115,34 @@ def create_steps(prefix: str, synth_data: str) -> list[ExecutorStep]:
                     attribute_filetype="vortex",
                     keep_if_missing=True,
                 ),
-                FilterConfig(
-                    type=FilterType.REMOVE_DOC,
-                    attribute_path=dedup_fuzzy_document_step.cd("data"),
-                    name="dup_doc",
-                    attribute_filetype="vortex",
-                    keep_if_missing=True,
-                ),
             ],
         ),
     )
 
-    ############################################################
     # Tokenize
-    tokenizer = "gpt2"
-
-    tokenize_config = TokenizeConfig(
-        train_paths=[consolidate_step],
-        validation_paths=[],
-        cache_path=this_output_path(),
-        tokenizer=tokenizer,
-    )
-
     tokenize_step = ExecutorStep(
         name=os.path.join(prefix, "tokenized"),
-        description=f"Tokenize raw text using the {tokenizer} tokenizer.",
         fn=tokenize,
-        config=tokenize_config,
+        config=TokenizeConfig(
+            train_paths=[consolidate_step],
+            validation_paths=[],
+            cache_path=this_output_path(),
+            tokenizer="gpt2",
+        ),
     )
 
-    # ############################################################
-    # Training
-    train_env_vars = {
-        "WANDB_API_KEY": "",
-        "WANDB_MODE": "disabled",
-        "JAX_TRACEBACK_FILTERING": "off",
-    }
-
-    pod_config = ResourceConfig.with_cpu()
-
+    # Train (tiny model for validation)
     train_step = ExecutorStep(
         name=os.path.join(prefix, "train"),
         fn=run_levanter_train_lm,
         config=TrainLmOnPodConfig(
             output_path=this_output_path(),
-            resources=pod_config,
-            env_vars=train_env_vars,
+            resources=ResourceConfig.with_cpu(),
+            env_vars={
+                "WANDB_API_KEY": "",
+                "WANDB_MODE": "disabled",
+                "JAX_TRACEBACK_FILTERING": "off",
+            },
             train_config=TrainLmConfig(
                 data=lm_data_config(tokenize_step),
                 hf_save_steps=1,
@@ -280,67 +159,130 @@ def create_steps(prefix: str, synth_data: str) -> list[ExecutorStep]:
         ),
     )
 
-    ##### Evaluate
-
-    # evaluate_step = evaluate_helm_on_step(train_step, ["mmlu"], max_eval_instances=10)
-
     return [
         transform_hq_data_step,
         transform_lq_data_step,
         dedup_exact_paragraph_step,
-        dedup_fuzzy_document_step,
-        validate_exact_dedup_step,
-        validate_fuzzy_dedup_step,
         consolidate_step,
         tokenize_step,
         train_step,
-        # evaluate_step,
     ]
 
 
-@draccus.wrap()
-def main(config: ExecutorMainConfig):
+# ---------------------------------------------------------------------------
+# S3 helpers
+# ---------------------------------------------------------------------------
+
+
+def _upload_tree(local_root: Path, s3_dest: str) -> None:
+    fs, _ = fsspec.core.url_to_fs(s3_dest)
+    for path in local_root.rglob("*"):
+        if not path.is_file():
+            continue
+        rel = path.relative_to(local_root)
+        fs.put(str(path), f"{s3_dest}/{rel}")
+
+
+def _rm_s3(s3_prefix: str) -> None:
+    fs, _ = fsspec.core.url_to_fs(s3_prefix)
     try:
-        if config.prefix is not None:
-            bucket_prefix = config.prefix
+        fs.rm(s3_prefix, recursive=True)
+    except FileNotFoundError:
+        pass
+
+
+def _s3_env_vars() -> dict[str, str]:
+    return {k: os.environ[k] for k in _S3_ENV_KEYS if k in os.environ}
+
+
+# ---------------------------------------------------------------------------
+# Executor entry point (runs inside the Iris job on remote clusters)
+# ---------------------------------------------------------------------------
+
+
+def _run_executor(prefix: str, synth_data: str) -> None:
+    config = ExecutorMainConfig(
+        prefix=prefix,
+        executor_info_base_path=f"{prefix}/experiments",
+    )
+    steps = create_steps("quickstart-tests", synth_data)
+    executor_main(config, steps=steps)
+
+
+# ---------------------------------------------------------------------------
+# Main
+# ---------------------------------------------------------------------------
+
+
+def main():
+    parser = argparse.ArgumentParser(description="Run full marin pipeline on Iris")
+    parser.add_argument("--controller-url", required=True)
+    args = parser.parse_args()
+
+    s3_base = os.environ.get("MARIN_CI_S3_PREFIX")
+
+    if s3_base:
+        run_id = f"marin-itest-{uuid.uuid4().hex[:8]}"
+        prefix = f"{s3_base}/{run_id}"
+        synth_data = f"{prefix}/quickstart-data"
+        logger.info("Uploading test fixtures to %s", synth_data)
+        _upload_tree(LOCAL_SYNTH_DATA, synth_data)
+        cleanup = lambda: _rm_s3(prefix)  # noqa: E731
+    else:
+        prefix = tempfile.mkdtemp(prefix="iris-marin-itest-")
+        synth_data = str(LOCAL_SYNTH_DATA)
+        cleanup = lambda: shutil.rmtree(prefix, ignore_errors=True)  # noqa: E731
+
+    os.environ["MARIN_PREFIX"] = prefix
+    os.environ["WANDB_MODE"] = "disabled"
+    os.environ["WANDB_API_KEY"] = ""
+    os.environ["JAX_TRACEBACK_FILTERING"] = "off"
+
+    try:
+        iris_client = FrayIrisClient(
+            controller_address=args.controller_url,
+            workspace=REPO_ROOT,
+        )
+
+        if s3_base:
+            logger.info("Submitting executor as Iris job (S3 mode)")
+            env_vars = {
+                "MARIN_PREFIX": prefix,
+                "WANDB_MODE": "disabled",
+                "WANDB_API_KEY": "",
+                "JAX_TRACEBACK_FILTERING": "off",
+                **_s3_env_vars(),
+            }
+
+            with set_current_client(iris_client):
+                handle = iris_client.submit(
+                    JobRequest(
+                        name=f"marin-itest-{uuid.uuid4().hex[:8]}",
+                        entrypoint=Entrypoint.from_callable(
+                            _run_executor,
+                            args=(prefix, synth_data),
+                        ),
+                        resources=ResourceConfig.with_cpu(),
+                        environment=create_environment(env_vars=env_vars),
+                    )
+                )
+                handle.wait(raise_on_failure=True, stream_logs=True)
         else:
-            bucket_prefix = "/tmp"  # Default to a temporary directory
+            logger.info("Running executor in-process (local mode)")
+            config = ExecutorMainConfig(
+                prefix=prefix,
+                executor_info_base_path=f"{prefix}/experiments",
+            )
+            steps = create_steps("quickstart-tests", synth_data)
+            with set_current_client(iris_client):
+                executor_main(config, steps=steps)
 
-        _setup_log_tee("/tmp/integration_test.log")
-
-        experiment_prefix = "quickstart-tests"
-        config = dataclasses.replace(
-            config, prefix=bucket_prefix, executor_info_base_path=os.path.join(bucket_prefix, "experiments")
-        )
-
-        os.environ["MARIN_PREFIX"] = bucket_prefix
-
-        # start Ray explicitly and set it as the current cluster
-        # N.B. This script must not be launched via `uv run`, or Ray will prefer to use `uv` for all execution
-        # ignoring package dependencies specified in each step.
-        if "uv run" in " ".join(sys.argv):
-            raise RuntimeError("integration_test.py must not be launched via `uv run`. Please run it directly.")
-        import ray
-
-        ray.init(
-            resources={"head_node": 1},
-            runtime_env={"working_dir": None},
-            num_cpus=os.cpu_count(),
-            _memory=1024 * 1024 * 1024 * 1024,  # 1TB
-        )
-        with set_current_client(RayClient()):
-            # path to synthetic test data
-            synth_data: str = "./tests/quickstart-data"
-            # delete all previous runs
-            if os.path.exists(os.path.join(bucket_prefix, experiment_prefix)):
-                os.system(f"rm -rf {os.path.join(bucket_prefix, experiment_prefix)}")
-            steps = create_steps(experiment_prefix, synth_data)
-            config = dataclasses.replace(config)
-            executor_main(config, steps=steps)
-            logger.info(f"Execution completed successfully. All outputs are in {bucket_prefix}/{experiment_prefix}")
-    except Exception as e:
-        logger.error(f"Error in main execution: {e}")
-        raise e
+        logger.info("Pipeline completed successfully")
+    except Exception:
+        logger.exception("Pipeline failed")
+        sys.exit(1)
+    finally:
+        cleanup()
 
 
 if __name__ == "__main__":
